requirements.txt
requests
matplotlib
obspy
cartopy
pyyaml
tqdm
